products as MVC, show farm points in map fetch all products from production points

+++ controller/productController.php
@@ -8,31 +8,93 @@ switch ($_SERVER['REQUEST_METHOD']) {
         if ((isset($_GET['productId'])) && $_GET['productId'] !== 0){
             $productId = $_GET['productId'];
             echo getProduct($productId);
+            break;
+        }
+
+        if ((isset($_GET['userId'])) && $_GET['userId'] !== 0){
+            $userId = $_GET['userId'];
+            echo fetchAllProducts($userId);
+            break;
+        }
+
+        if ((isset($_GET['productionLocationId'])) && $_GET['productionLocationId'] !== 0){
+            $productionLocation = $_GET['productionLocationId'];
+            echo fetchAllProductsFromLocation($productionLocation);
+            break;
         }
         break;
-        
+
     case 'POST':
-        
+        if (isset($_POST['userId']) && isset($_POST['fetchAllProducts'])){
+            $userId = $_POST['userId'];
+            echo fetchAllProducts($userId);
+            break;
+        }
+
         $product = new product($_POST);
 
         if ($product->isDelete == true){
             echo deleteProduct($product->productId);
+            break;
         } else{
             if ($product->productId == 0){
-            addProduct($product->productName, $product->productDesc, $product->productCategory, $product->productionLocation, $product->isProcessedProduct, $product->isAvailable, $product->pricePerUnit, $product->quantityOfPrice, $product->unit, $product->productRating, $files, $product->productFeatures);
-        }else{
-            updateProducts($product->productId,$product->productName, $product->productDesc, $product->productCategory, $product->productionLocation, $product->isProcessedProduct, $product->isAvailable, $product->pricePerUnit, $product->quantityOfPrice, $product->unit, $product->productRating);
+                addProduct($product->productName, $product->productDesc, $product->productCategory, $product->productionLocation, $product->isProcessedProduct, $product->isAvailable, $product->pricePerUnit, $product->quantityOfPrice, $product->unit, $product->productRating, $files, $product->productFeatures, $product->productFeaturesId);
+                break;
+            }else{
+                updateProducts($product->productId,$product->productName, $product->productDesc, $product->productCategory, $product->productionLocation, $product->isProcessedProduct, $product->isAvailable, $product->pricePerUnit, $product->quantityOfPrice, $product->unit, $product->productRating, $product->productFeatures, $product->productFeaturesId);
+                break;
+            }
         }
-        }
-        
+
         break;
+}
+
+
+
+function fetchAllProducts($userId){
+    $products = [];
+    global $dbConnection;
+    $fetchProductQuery = "SELECT * FROM products p
+        LEFT JOIN images i on (i.entity_id = p.product_id and i.image_type = 1)
+        WHERE p.producer_id = '$userId'
+        GROUP BY p.product_id ORDER BY p.created_date DESC;";
+
+
+    //    $fetchProductFeaturesQuery = "SELECT * FROM product_feature p WHERE p.product_id in (SELECT product_id FROM products pd where pd.producer_id = $userId);";
+    //
+    //    $fetchProductQuery .= $fetchProductFeaturesQuery;
+
+
+
+
+    $productData = [];
+
+
+    if ($result = mysqli_query($dbConnection, $fetchProductQuery)) {
+        while ($row = mysqli_fetch_assoc($result)) {
+            $tempProductData = $row;
+            $productId = $row['product_id'];
+            $fetchProductFeaturesQuery = "SELECT * FROM product_feature p WHERE p.product_id = $productId;";
+            $featureResult = mysqli_query($dbConnection, $fetchProductFeaturesQuery) or die(mysql_error());
+            $featureArray = [];
+            while($featureRow = mysqli_fetch_assoc($featureResult)){
+                array_push($featureArray, $featureRow);
+            }
+            $tempProductData['product_feature'] = $featureArray;
+            array_push($productData,$tempProductData);
+        }
+
+    }
+
+    mysqli_close($dbConnection);
+    return json_encode($productData);
 }
 
 
 
 
 //Add product details to product table
-function addProduct($productName, $productDescription, $productCategory, $productionLocation, $isProcessedProduct, $isAvailable, $productPrice, $priceQuantity, $unit, $productRating, $fileNameArray, $productFeaturesArray){
+function addProduct($productName, $productDescription, $productCategory, $productionLocation, $isProcessedProduct, $isAvailable, $productPrice, $priceQuantity, $unit, $productRating, $fileNameArray, $productFeaturesArray, $productFeatureIdArray){
     global $dbConnection;
     /* Start transaction */
     mysqli_begin_transaction($dbConnection);
@@ -53,20 +115,20 @@ function addProduct($productName, $productDescription, $productCategory, $produc
 
 
             $productFeaturesCount = count($productFeaturesArray);
-            $productfeatureQuery = "INSERT INTO product_feature (product_id, feature_type) VALUES ";
+            $productfeatureQuery = PrepareFeatureQuery($productFeaturesArray, $productFeatureIdArray, $productId);//"INSERT INTO product_feature (product_id, feature_type) VALUES ";
             if ($productFeaturesCount > 0){
-                for ($i = 0; $i<$productFeaturesCount; $i++){
-                    $featureType = $productFeaturesArray[$i];
-                    $productfeatureQuery .= "($productId, $featureType)";
-                    if ($i===$productFeaturesCount-1){
-                        $productfeatureQuery .= ";";
-                    }else{
-                        $productfeatureQuery .= ", ";
-                    }
-                }
-                echo $productImageQuery;
-                if (mysqli_query($dbConnection, $productfeatureQuery)){
-                    echo "  Files inserted succesfully   ";
+                //                for ($i = 0; $i<$productFeaturesCount; $i++){
+                //                    $featureType = $productFeaturesArray[$i];
+                //                    $productfeatureQuery .= "($productId, $featureType)";
+                //                    if ($i===$productFeaturesCount-1){
+                //                        $productfeatureQuery .= ";";
+                //                    }else{
+                //                        $productfeatureQuery .= ", ";
+                //                    }
+                //                }
+                //                echo $productfeatureQuery;
+                if (mysqli_multi_query($dbConnection, $productfeatureQuery)){
+                    echo "  features inserted succesfully   ";
                     //                    mysqli_commit($dbConnection);
                 }else{
                     echo "product creation failed at inserting images,";
@@ -115,7 +177,7 @@ function addProduct($productName, $productDescription, $productCategory, $produc
 
 
 
-function updateProducts($productId, $productName, $productDesc, $productCategory, $productionLocation, $isProcessedProduct, $isAvailable, $pricePerUnit, $quantityOfPrice, $unit, $productRating){
+function updateProducts($productId, $productName, $productDesc, $productCategory, $productionLocation, $isProcessedProduct, $isAvailable, $pricePerUnit, $quantityOfPrice, $unit, $productRating, $productFeatures, $featureIdArray){
 
     ob_start();
     global $dbConnection;
@@ -130,20 +192,26 @@ function updateProducts($productId, $productName, $productDesc, $productCategory
         $updateProductQuery .= "SET product_name = '$productName', product_description = '$productDesc', product_category = $productCategory, production_location = $productionLocation, is_processed_product = $isProcessedProduct, is_available = $isAvailable, price_per_unit = $pricePerUnit, quantity_of_price = $quantityOfPrice, unit = $unit, product_rating = $productRating ";
         $updateProductQuery .= "WHERE product_id = $productId AND producer_id = $producerId;";
 
-        mysqli_begin_transaction($dbConnection);
+
+        $productfeatureQuery = PrepareFeatureQuery($productFeatures, $featureIdArray, $productId);
+
+        $updateProductQuery .= $productfeatureQuery;
+
+
+        //        mysqli_begin_transaction($dbConnection);
 
         try{
-            echo "<script>console.log('PHP: " . $updateProductQuery . "');</script>";
-            if (mysqli_query($dbConnection, $updateProductQuery)){
-                mysqli_commit($dbConnection);
+            //            echo "<script>console.log('PHP: " . $updateProductQuery . "');</script>";
+            if (mysqli_multi_query($dbConnection, $updateProductQuery)){
+                //                mysqli_commit($dbConnection);
                 echo "<script>console.log('PHP: Successfully updated prodcut details');</script>";
             }else{
                 echo "<script>console.log('PHP: faield to update prodcut details');</script>";
-                mysqli_rollback($dbConnection);
+                //                mysqli_rollback($dbConnection);
             }
         }catch(mysqli_sql_exception $exception){
             echo "<script>console.log('PHP: faield to update prodcut details exception');</script>";
-            mysqli_rollback($dbConnection);
+            //            mysqli_rollback($dbConnection);
             var_dump($exception);
             throw $exception;
 
@@ -159,20 +227,50 @@ function getProduct($productId){
     if (isTokenValid()){
         $producerId = $_SESSION["userId"];
         $getProductQuery = "SELECT * FROM products p ";
-//        $getProductQuery .= "JOIN product_feature pf on pf.product_id = p.product_id ";
+        //        $getProductQuery .= "JOIN product_feature pf on pf.product_id = p.product_id ";
         $getProductQuery .= "WHERE p.product_id = $productId AND p.producer_id = $producerId;";
-        $getProductResult = mysqli_query($dbConnection, $getProductQuery);
-        confirmQuery($getProductResult);
-        if ($getProductResult->num_rows > 0){
-            $result = mysqli_fetch_all($getProductResult, MYSQLI_ASSOC);
-            ob_end_clean();
-            return json_encode($result[0]);
-        }else{
-            echo "<script>console.log('PHP: No products found with the given product id');</script>";
+
+        $getProductQuery .= "SELECT * FROM product_feature pf ";
+        $getProductQuery .= "WHERE pf.product_id = $productId;";
+        $productData = [];
+        if (mysqli_multi_query($dbConnection, $getProductQuery)) {
+            do {
+                // Store first result set
+                if ($result = mysqli_store_result($dbConnection)) {
+                    while ($row = $result->fetch_all(MYSQLI_ASSOC)) {
+                        array_push ($productData, $row);
+                    }
+                    //                    mysqli_free_result($result);
+                }
+
+            } while(mysqli_more_results($dbConnection) && mysqli_next_result($dbConnection));
+
         }
+
+        mysqli_close($dbConnection);
+        return json_encode($productData);
+
     }else{
         echo "<script>console.log('PHP: Authentication Failed');</script>";
     }
+}
+
+function PrepareFeatureQuery($featureArray, $featureIdArray, $productId){
+    $featureQuery = "";
+
+    for ($i = 0; $i < count($featureIdArray); $i++){
+        $featureType = $featureArray[$i];
+        $featureId = $featureIdArray[$i];
+        if ($featureId == 0 && $featureType != 0){
+            $featureQuery .= "INSERT INTO product_feature (product_id, feature_type) VALUES ";
+            $featureQuery .= "($productId, $featureType);";
+        }elseif($featureId != 0 && $featureType != 0){
+            $featureQuery .= "UPDATE product_feature  SET product_id = $productId, feature_type = $featureType WHERE id = $featureId;";
+        }elseif($featureId != 0 && $featureType == 0){
+            $featureQuery .= "DELETE FROM product_feature WHERE id = $featureId;";
+        }
+    }
+    return $featureQuery;
 }
 
 
@@ -196,4 +294,25 @@ function deleteProduct($productId){
     }
     return false;
 }
+
+
+function fetchAllProductsFromLocation($locationId){
+    ob_start();
+    global $dbConnection;
+
+    $productsQuery = "SELECT * FROM products p where p.production_location = $locationId;";
+    $productData = [];
+    if ($result = mysqli_query($dbConnection, $productsQuery)) {
+        while ($row = $result->fetch_all(MYSQLI_ASSOC)) {
+            array_push ($productData, $row);
+        }
+    }
+
+    mysqli_close($dbConnection);
+    return json_encode($productData);
+
+
+
+}
+
 ?>
